requirements.txt
flask
numpy
redis
imutils
Pillow
opencv-contrib-python
Flask-SQLAlchemy
Flask-Migrate
psycopg2
Flask-SocketIO
Flask-Cors
gunicorn
eventlet==0.30.2
gevent
gevent-websocket
boto3
botocore
moto==1.3.13